Revert "Rename to gcache (#46)"
This reverts commit 06bec7e8691e3dce020278d373e7a58178ca3e38.
