+++ tasks.py
@@ -39,7 +39,7 @@ def type_check(ctx: Context) -> None:
     ctx : Context
         Invoke context.
     """
-    ctx.run("poetry run mypy --package gcache --namespace-packages", **COMMON_PARAMS)
+    ctx.run("poetry run mypy --package cachegalileo --namespace-packages", **COMMON_PARAMS)
     ctx.run("poetry run mypy --package tests --namespace-packages", **COMMON_PARAMS)
 
 
@@ -58,4 +58,4 @@ def setup_pre_commit(ctx: Context) -> None:
 
 @task
 def test(ctx: Context) -> None:
-    ctx.run("poetry run pytest -vvv --cov=gcache --cov-report=xml", **COMMON_PARAMS)
+    ctx.run("poetry run pytest -vvv --cov=cachegalileo --cov-report=xml", **COMMON_PARAMS)
